Move instance manager outside core.
